requirements.txt
Flask==3.0.0
qrcode==7.4.2
Pillow==10.2.0
flask-cors==4.0.0
gunicorn==21.2.0